Refactor grid properties
