+++ packages/perspective-viewer-hypergrid/src/js/hypergrid.js
@@ -14,114 +14,16 @@ const groupedHeaderPlugin = require("fin-hypergrid-grouped-header-plugin");
 const perspectivePlugin = require("./perspective-plugin");
 const PerspectiveDataModel = require("./PerspectiveDataModel");
 const {psp2hypergrid, page2hypergrid} = require("./psp-to-hypergrid");
-const {cloneDeep} = require("lodash");
 
 import {bindTemplate} from "@finos/perspective-viewer/dist/esm/utils.js";
 
 const TEMPLATE = require("../html/hypergrid.html");
 
 import style from "../less/hypergrid.less";
-import {get_styles, clear_styles} from "./styles.js";
+import {get_styles, clear_styles, default_grid_properties} from "./styles.js";
 import {set_formatters} from "./formatters.js";
 import {set_editors} from "./editors.js";
 import {treeLineRendererPaint} from "./hypergrid-tree-cell-renderer";
-
-const COLUMN_HEADER_FONT = "12px Helvetica, sans-serif";
-const GROUP_LABEL_FONT = "12px Open Sans, sans-serif";
-
-const base_grid_properties = {
-    autoSelectRows: false,
-    cellPadding: 5,
-    cellSelection: false,
-    columnSelection: false,
-    rowSelection: false,
-    checkboxOnlyRowSelections: false,
-    columnClip: true,
-    columnHeaderFont: COLUMN_HEADER_FONT,
-    columnHeaderForegroundSelectionFont: '12px "Arial", Helvetica, sans-serif',
-    columnsReorderable: false,
-    defaultRowHeight: 24,
-    editable: true,
-    editOnKeydown: true,
-    editorActivationKeys: ["alt", "esc"],
-    enableContinuousRepaint: false,
-    fixedColumnCount: 0,
-    fixedRowCount: 0,
-    fixedLinesHWidth: 1,
-    fixedLinesVWidth: 1,
-    font: '12px "Arial", Helvetica, sans-serif',
-    foregroundSelectionFont: '12px "Arial", Helvetica, sans-serif',
-    gridLinesH: false,
-    gridLinesV: true, // except: due to groupedHeaderPlugin's `clipRuleLines: true` option, only header row displays these lines
-    gridLinesUserDataArea: false, // restricts vertical rule line rendering to header row only
-    halign: "left",
-    headerTextWrapping: false,
-    hoverColumnHighlight: {enabled: false},
-    hoverRowHighlight: {
-        enabled: true,
-        backgroundColor: "#555"
-    },
-    hoverCellHighlight: {
-        enabled: true,
-        backgroundColor: "#333"
-    },
-    noDataMessage: "",
-    minimumColumnWidth: 50,
-    multipleSelections: false,
-    renderFalsy: false,
-    rowHeaderFont: "12px Arial, Helvetica, sans-serif",
-    rowHeaderForegroundSelectionFont: '12px "Arial", Helvetica, sans-serif',
-    rowResize: true,
-    scrollbarHoverOff: "visible",
-    rowHeaderCheckboxes: false,
-    rowHeaderNumbers: false,
-    showFilterRow: true,
-    showHeaderRow: true,
-    showTreeColumn: false,
-    singleRowSelectionMode: false,
-    sortColumns: [],
-    sortOnDoubleClick: true,
-    treeRenderer: "TreeCell",
-    treeHeaderFont: "12px Arial, Helvetica, sans-serif",
-    treeHeaderForegroundSelectionFont: '12px "Arial", Helvetica, sans-serif',
-    useBitBlit: false,
-    vScrollbarClassPrefix: "",
-    voffset: 0
-};
-
-const light_theme_overrides = {
-    backgroundColor: "#ffffff",
-    color: "#666",
-    lineColor: "#AAA",
-    // font: '12px Arial, Helvetica, sans-serif',
-    font: '12px "Open Sans", Helvetica, sans-serif',
-    foregroundSelectionFont: "12px amplitude-regular, Helvetica, sans-serif",
-    foregroundSelectionColor: "#666",
-    backgroundSelectionColor: "rgba(162, 183, 206, 0.3)",
-    selectionRegionOutlineColor: "rgb(45, 64, 85)",
-    columnHeaderColor: "#666",
-    columnHeaderHalign: "left", // except: group header labels always 'center'; numbers always 'right' per `setPSP`
-    columnHeaderBackgroundColor: "#fff",
-    columnHeaderForegroundSelectionColor: "#333",
-    columnHeaderBackgroundSelectionColor: "#40536d",
-    rowHeaderForegroundSelectionFont: "12px Arial, Helvetica, sans-serif",
-    treeHeaderColor: "#666",
-    treeHeaderBackgroundColor: "#fff",
-    treeHeaderForegroundSelectionColor: "#333",
-    treeHeaderBackgroundSelectionColor: "#40536d",
-    hoverCellHighlight: {
-        enabled: true,
-        backgroundColor: "#eeeeee"
-    },
-    hoverRowHighlight: {
-        enabled: true,
-        backgroundColor: "#f6f6f6"
-    }
-};
-
-function generateGridProperties(overrides) {
-    return Object.assign({}, cloneDeep(base_grid_properties), cloneDeep(overrides));
-}
 
 bindTemplate(TEMPLATE, style)(
     class HypergridElement extends HTMLElement {
@@ -144,22 +46,8 @@ bindTemplate(TEMPLATE, style)(
                 host.removeAttribute("hidden");
                 this.grid.get_styles = () => get_styles(this);
 
-                this.grid.installPlugins([
-                    perspectivePlugin,
-                    [
-                        groupedHeaderPlugin,
-                        {
-                            paintBackground: null, // no group header label decoration
-                            columnHeaderLines: false, // only draw vertical rule lines between group labels
-                            groupConfig: [
-                                {
-                                    halign: "center", // center group labels
-                                    font: GROUP_LABEL_FONT
-                                }
-                            ]
-                        }
-                    ]
-                ]);
+                const grid_properties = default_grid_properties();
+                this.grid.installPlugins([perspectivePlugin, [groupedHeaderPlugin, grid_properties.groupedHeader]]);
 
                 // Broken in fin-hypergrid-grouped-header 0.1.2
                 let _old_paint = this.grid.cellRenderers.items.GroupedHeader.paint;
@@ -168,9 +56,6 @@ bindTemplate(TEMPLATE, style)(
                     return _old_paint.call(this, gc, config);
                 };
 
-                const grid_properties = generateGridProperties(light_theme_overrides);
-
-                grid_properties["showRowNumbers"] = grid_properties["showCheckboxes"] || grid_properties["showRowNumbers"];
                 this.grid.addProperties(grid_properties);
                 const styles = get_styles(this);
                 this.grid.addProperties(styles[""]);
